etcdctl: add a backup command
This current version is a hack but I want to get a check on the general
idea. The next piece is to the new singleton member to the memberstore
configured to advertise localhost. I don't know how to go from this
configuration to something a user can setup a distributed cluster on
though yet.